[ASTERIXDB-3392] Fix format in tests
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/18778
Tested-by: Wail Alkowaileet <wael.y.k@gmail.com>
Integration-Tests: Wail Alkowaileet <wael.y.k@gmail.com>

+++ asterixdb/asterix-external-data/src/main/java/org/apache/asterix/external/writer/printer/parquet/AsterixParquetTypeMap.java
@@ -31,7 +31,7 @@ public class AsterixParquetTypeMap {
                     Map.entry(ATypeTag.STRING, PrimitiveType.PrimitiveTypeName.BINARY),
                     Map.entry(ATypeTag.TINYINT, PrimitiveType.PrimitiveTypeName.INT32),
                     Map.entry(ATypeTag.SMALLINT, PrimitiveType.PrimitiveTypeName.INT32),
-                    Map.entry(ATypeTag.INTEGER, PrimitiveType.PrimitiveTypeName.INT64),
+                    Map.entry(ATypeTag.INTEGER, PrimitiveType.PrimitiveTypeName.INT32),
                     Map.entry(ATypeTag.BIGINT, PrimitiveType.PrimitiveTypeName.INT64),
                     Map.entry(ATypeTag.FLOAT, PrimitiveType.PrimitiveTypeName.FLOAT),
                     Map.entry(ATypeTag.DOUBLE, PrimitiveType.PrimitiveTypeName.DOUBLE),
